refactor of core traits
