+++ src/Concerns/CanTranslate.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace atk4\core\Concerns;
+
+use atk4\core\Translator\Translator;
+
+/**
+ * If a class use this trait, string can be translated calling method translate.
+ */
+trait CanTranslate
+{
+    /**
+     * Translates the given message.
+     *
+     * @param string      $message    The message to be translated
+     * @param array       $parameters Array of parameters used to translate message
+     * @param string|null $context    The domain for the message or null to use the default
+     * @param string|null $locale     The locale or null to use the default
+     *
+     * @return string The translated string
+     */
+    public function _($message, array $parameters = [], ?string $domain = null, ?string $locale = null): string
+    {
+        // if App is present
+        if (isset($this->app) && method_exists($this->app, '_')) {
+            return $this->app->_($message, $parameters, $domain, $locale);
+        }
+
+        return Translator::instance()->_($message, $parameters, $domain, $locale);
+    }
+}
