+++ src/Concerns/HasName.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace atk4\core\Concerns;
+
+/**
+ * This trait makes it possible to set name of your object.
+ */
+trait HasName
+{
+    /**
+     * Unique object name.
+     *
+     * @var string
+     */
+    public $name;
+}
