+++ src/NameTrait.php
@@ -3,21 +3,19 @@
 namespace atk4\core;
 
 /**
- * This trait makes it possible to set name of your object.
+ * @deprecated use HasName instead.
  */
 trait NameTrait
 {
+    use Concerns\HasName;
+    
     /**
      * Check this property to see if trait is present in the object.
      *
      * @var bool
+     * 
+     * @deprecated use automated trait detection with CanCheckTraits
+     * 
      */
     public $_nameTrait = true;
-
-    /**
-     * Unique object name.
-     *
-     * @var string
-     */
-    public $name;
 }
